Digital_Ocean_backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
pydantic-settings
python-dotenv
httpx
